신규 기능 [202306221830] - 플래너 모달 > insert 연결
신규 파일 [202306221830] - LocalDAO, LocalDAOImpl 추가

결함 [202306222040] - 플래너 Mapper list 쿼리 오류 수정

개선
    [202306222000] - 플래너 ScheduleDTO y,x 타입 변경
    [202306222100] - 플래너 list 검색옵션 수정
    [202306222150] - 플래너 insert - 작업중

+++ trakker/src/main/java/com/example/trakker/model/planner/dto/ScheduleDTO.java
@@ -9,6 +9,6 @@ public class ScheduleDTO {
     private Integer sNum; //날짜별 일정번호
     private String sPoint; //장소명
     private String sMemo; //장소 메모
-    private Integer x;
-    private Integer y;
+    private Float x;
+    private Float y;
 }
